requirements.txt
Flask==2.3.2
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.0
python-whois==0.8.0
joblib==1.3.2
urllib3==2.0.7
